// src/app/models/vehicle.ts
export interface Vehicle {
  'id': number;
  'name': string;
  'model': string;
  'manufacturer': string;
  'cost_in_credits': string;
  'length': string;
  'max_atmosphering_speed': string;
  'crew': string;
  'passengers': string;
  'cargo_capacity': string;
  'consumables': string;
  'vehicle_class': string;
  'pilots': [];
  'films': [];
  'created': string;
  'edited': string;
  'url': string;
}